大类型向小类型转换需要显示转换

+++ src/chapter02/Example01.java
@@ -3,7 +3,8 @@ package chapter02;
 public class Example01 {
 	public static void main(String[] args) {
 		int num = 4;
-		byte b = (byte) num;
+		//大类型向小类型转换，需要显示转换
+		byte b = num;
 		System.out.println(b);
 	}
 }
